Fuse3

+++ query.c
@@ -19,7 +19,7 @@
 #include <fcntl.h>
 #include <time.h>
 #include <libgen.h>
-#include <fuse/fuse.h>
+#include <fuse.h>
 #ifdef HAVE_MYSQL_MYSQL_H
 #include <mysql/mysql.h>
 #endif
@@ -88,7 +88,7 @@ int query_getattr(MYSQL *mysql, const char *path, struct stat *stbuf)
       return ret;
 
     snprintf(sql, SQL_MAX,
-             "SELECT inode, mode, uid, gid, ctime, atime, mtime "
+             "SELECT inode, mode, uid, gid, ctime, atime, mtime, size "
              "FROM inodes WHERE inode=%ld",
              inode);
 
@@ -114,6 +114,7 @@ int query_getattr(MYSQL *mysql, const char *path, struct stat *stbuf)
     }
     row = mysql_fetch_row(result);
     if(!row){
+        log_printf(LOG_ERROR, "ERROR: mysql_fetch_row()\n");
         return -EIO;
     }
 
@@ -124,6 +125,7 @@ int query_getattr(MYSQL *mysql, const char *path, struct stat *stbuf)
     stbuf->st_ctime = atol(row[4]);
     stbuf->st_atime = atol(row[5]);
     stbuf->st_mtime = atol(row[6]);
+    stbuf->st_size = atol(row[7]);
     stbuf->st_nlink = nlinks;
 
     mysql_free_result(result);
@@ -138,7 +140,7 @@ int query_getattr(MYSQL *mysql, const char *path, struct stat *stbuf)
  * good testcase :)
  *
  * If any of the name, inode, parent, or nlinks are given, those values will be
- * recorded form the inode data to the given buffers.  The name is written to
+ * recorded from the inode data to the given buffers.  The name is written to
  * the given name_len.
  *
  * @return 0 if successful
@@ -156,16 +158,16 @@ int query_inode_full(MYSQL *mysql, const char *path, char *name, size_t name_len
 		      long *inode, long *parent, long *nlinks)
 {
     long ret;
-    char sql[SQL_MAX];
+    char sql[SQL_MAX*4];
     MYSQL_RES* result;
     MYSQL_ROW row;
 
     int depth = 0;
     char *pathptr = strdup(path), *pathptr_saved = pathptr;
     char *nameptr, *saveptr = NULL;
-    char sql_from[SQL_MAX], sql_where[SQL_MAX];
+    char sql_from[SQL_MAX/3], sql_where[SQL_MAX/3];
     char *sql_from_end = sql_from, *sql_where_end = sql_where;
-    char esc_name[PATH_MAX * 2];
+    char esc_name[PATH_MAX];
 
     // TODO: Handle too long or too nested paths that don't fit in SQL_MAX!!!
     sql_from_end += snprintf(sql_from_end, SQL_MAX, "tree AS t0");
@@ -174,6 +176,11 @@ int query_inode_full(MYSQL *mysql, const char *path, char *name, size_t name_len
         if (depth++ == 0) {
 	  pathptr = NULL;
 	}
+
+        if (strlen(nameptr) > 255) {
+            free(pathptr_saved);
+            return -ENAMETOOLONG;
+        }
 
         mysql_real_escape_string(mysql, esc_name, nameptr, strlen(nameptr));
 	sql_from_end += snprintf(sql_from_end, SQL_MAX, " LEFT JOIN tree AS t%d ON t%d.inode = t%d.parent",
@@ -239,7 +246,7 @@ int query_inode_full(MYSQL *mysql, const char *path, char *name, size_t name_len
  * a nestable return value.
  *
  * @return ID of inode
- * @return < 0 result of query_inode_full() if that ufnction reports a failure
+ * @return < 0 result of query_inode_full() if that function reports a failure
  * @param mysql handle to connection to the database
  * @param path (full) pathname of inode to find
  */
@@ -247,6 +254,8 @@ long query_inode(MYSQL *mysql, const char *path)
 {
     long inode, ret;
     
+    if (strlen(path) > PATH_MAX)
+        return -ENAMETOOLONG;
     ret = query_inode_full(mysql, path, NULL, 0, &inode, NULL, NULL);
     if (ret < 0)
       return ret;
@@ -264,20 +273,16 @@ long query_inode(MYSQL *mysql, const char *path)
  *
  * @return 0 on success; non-zero return of mysql_query() on error
  * @param mysql handle to connection to the database
- * @param path pathname of file to truncate
+ * @param inode node to operate on
  * @param length new length of file
  */
-int query_truncate(MYSQL *mysql, const char *path, off_t length)
+int query_truncate(MYSQL *mysql, long inode, off_t length)
 {
     int ret;
     char sql[SQL_MAX];
     struct data_blocks_info info;
 
     fill_data_blocks_info(&info, length, 0);
-
-    long inode = query_inode(mysql, path);
-    if (inode < 0)
-      return inode;
 
     lock_inode(mysql, inode);
 
@@ -295,7 +300,7 @@ int query_truncate(MYSQL *mysql, const char *path, off_t length)
     if ((ret = mysql_query(mysql, sql))) goto err_out;
 
     snprintf(sql, SQL_MAX,
-             "UPDATE inodes SET size=%lld WHERE inode=%ld",
+             "UPDATE inodes SET size=%lld, mtime=UNIX_TIMESTAMP(NOW()), ctime=UNIX_TIMESTAMP(NOW()) WHERE inode=%ld",
              (long long)length, inode);
     log_printf(LOG_D_SQL, "sql=%s\n", sql);
     if ((ret = mysql_query(mysql, sql))) goto err_out;
@@ -327,6 +332,10 @@ int query_mkdirentry(MYSQL *mysql, long inode, const char *name, long parent)
     char sql[SQL_MAX];
     char esc_name[PATH_MAX * 2];
 
+    /* 
+     * Should really update ctime in inode --- but if we did that we could 
+     * add an nlinks count in the inode relation and get rid of the funcky join
+     */
     mysql_real_escape_string(mysql, esc_name, name, strlen(name));
     snprintf(sql, SQL_MAX,
              "INSERT INTO tree (name, parent, inode) VALUES ('%s', %ld, %ld)",
@@ -343,7 +352,10 @@ int query_mkdirentry(MYSQL *mysql, long inode, const char *name, long parent)
 }
 
 /**
- * The opposite of query_mkdirentry(), this function deletes a directory from the tree with a parent that matches the inode given.
+ * The opposite of query_mkdirentry(), this function deletes a link
+ * from the tree with a parent that matches the inode given.
+ * It will refuse to delete an entry if it has children (i.e. is  a 
+ * non-empty directory)
  *
  * @return 0 if successful
  * @return -EIO if the result of mysql_query() is non-zero
@@ -351,12 +363,31 @@ int query_mkdirentry(MYSQL *mysql, long inode, const char *name, long parent)
  * @param name name (relative)  of directory to delete
  * @param parent inode of directory holding the directory
  */
-int query_rmdirentry(MYSQL *mysql, const char *name, long parent)
+int query_rmdirentry(MYSQL *mysql, const char *name, long inode, long parent)
 {
     int ret;
+    MYSQL_RES *result;
     char sql[SQL_MAX];
     char esc_name[PATH_MAX * 2];
 
+    snprintf(sql, SQL_MAX, "select inode from tree where parent = %ld\n", inode);
+    ret = mysql_query(mysql, sql);
+    if(ret) {
+        log_printf(LOG_ERROR, "ERROR: mysql_query()\n");
+        log_printf(LOG_ERROR, "mysql_error: %s\n", mysql_error(mysql));
+        return -EIO;
+    }
+
+    result = mysql_store_result(mysql);
+    if(!result){
+        log_printf(LOG_ERROR, "ERROR: mysql_store_result()\n");
+        log_printf(LOG_ERROR, "mysql_error: %s\n", mysql_error(mysql));
+        return -EIO;
+    }
+
+    if (mysql_num_rows(result) != 0)
+        return -ENOTEMPTY;
+    
     mysql_real_escape_string(mysql, esc_name, name, strlen(name));
     snprintf(sql, SQL_MAX,
              "DELETE FROM tree WHERE name='%s' AND parent=%ld",
@@ -467,9 +498,6 @@ long query_mkdir(MYSQL *mysql, const char *path, mode_t mode, long parent)
  * The set of results is not ordered, so results would be in the "natural order"
  * of the database.
  *
- * for the kernel's implementation of a chmod() call in an inode on the FUSE
- * filesystem.
- *
  * @see http://linux.die.net/man/2/readdir
  *
  * @return 0 on success; -EIO on failure (non-zero return from mysql_query() function)
@@ -478,14 +506,15 @@ long query_mkdir(MYSQL *mysql, const char *path, mode_t mode, long parent)
  * @param buf buffer to pass to filler function
  * @param filler fuse_fill_dir_t function-pointer used to process each directory entry
  */
-int query_readdir(MYSQL *mysql, long inode, void *buf, fuse_fill_dir_t filler)
+int query_readdir(MYSQL *mysql, long inode, void *buf, fuse_fill_dir_t filler, int flag)
 {
     int ret;
     char sql[SQL_MAX];
     MYSQL_RES* result;
     MYSQL_ROW row;
+    struct stat st;
 
-    snprintf(sql, sizeof(sql), "SELECT name FROM tree WHERE parent = '%ld'",
+    snprintf(sql, sizeof(sql), "SELECT tree.name, tree.inode, inodes.mode FROM tree inner join inodes on tree.inode = inodes.inode WHERE tree.parent = '%ld'",
              inode);
 
     ret = mysql_query(mysql, sql);
@@ -500,13 +529,16 @@ int query_readdir(MYSQL *mysql, long inode, void *buf, fuse_fill_dir_t filler)
         return -EIO;
     }
 
+    memset(&st, 0, sizeof st);
     while((row = mysql_fetch_row(result)) != NULL){
-        filler(buf, (char*)basename(row[0]), NULL, 0);
+        st.st_ino = atol(row[1]);
+        st.st_mode = atol(row[2]);
+        filler(buf, (char*)basename(row[0]), &st, 0, 0);
     }
 
     mysql_free_result(result);
 
-    return ret;
+    return 0;
 }
 
 /**
@@ -561,7 +593,7 @@ int query_chown(MYSQL *mysql, long inode, uid_t uid, gid_t gid)
     char sql[SQL_MAX];
     size_t index;
 
-    index = snprintf(sql, SQL_MAX, "UPDATE inodes SET ");
+    index = snprintf(sql, SQL_MAX, "UPDATE inodes SET ctime=UNIX_TIMESTAMP(NOW()),");
     if (uid != (uid_t)-1)
     	index += snprintf(sql + index, SQL_MAX - index, 
 			  "uid=%d ", uid);
@@ -597,7 +629,7 @@ int query_chown(MYSQL *mysql, long inode, uid_t uid, gid_t gid)
  * @param inode inode to update the atime, mtime
  * @param time utimbuf with new actime, modtime, to set into access and modification times
  */
-int query_utime(MYSQL *mysql, long inode, struct utimbuf *time)
+int query_utime(MYSQL *mysql, long inode, const struct timespec tv[2])
 {
     int ret;
     char sql[SQL_MAX];
@@ -606,7 +638,7 @@ int query_utime(MYSQL *mysql, long inode, struct utimbuf *time)
              "UPDATE inodes "
              "SET atime=%ld, mtime=%ld "
              "WHERE inode=%lu",
-             time->actime, time->modtime, inode);
+             tv[0].tv_sec, tv[1].tv_sec, inode);
 
     log_printf(LOG_D_SQL, "sql=%s\n", sql);
 
@@ -965,17 +997,19 @@ ssize_t query_size(MYSQL *mysql, long inode)
 
     if(mysql_num_rows(result) != 1 || mysql_num_fields(result) != 1){
         mysql_free_result(result);
+        log_printf(LOG_ERROR, "ERROR: non-unique number of rows for %d\n", inode);
         return -EIO;
     }
 
     row = mysql_fetch_row(result);
     if(!row){
+        log_printf(LOG_ERROR, "ERROR: row-fetch failed for %d\n", inode);
         return -EIO;
     }
 
-    if(row[0]){
+    if (row[0]) {
         ret = atoll(row[0]);
-    }else{
+    } else {
         ret = 0;
     }
     mysql_free_result(result);
@@ -988,7 +1022,7 @@ ssize_t query_size(MYSQL *mysql, long inode)
  *
  * @return -ENXIO if the inode/seq pair is not found (zero rows returned, implying that block doesn't exist)
  * @return -EIO if no row is returned (implying an error in the query response, signaled by mysql_fetch_row() returning NULL)
- * @return 0 if the rown is NULL (implying no result?)
+ * @return 0 if the row is NULL (implying no result?)
  * @return 1 - DATA_BLOCK_SIZE (size of the actual block)
  * @param mysql handle to connection to the database
  * @param inode inode of the file in question
@@ -1025,6 +1059,7 @@ ssize_t query_size_block(MYSQL *mysql, long inode, unsigned long seq)
 
     row = mysql_fetch_row(result);
     if(!row){
+        log_printf(LOG_ERROR, "Could not fetch row: query_size_block\n");
         return -EIO;
     }
 
@@ -1056,6 +1091,16 @@ int query_rename(MYSQL *mysql, const char *from, const char *to)
     char *tmp, *new_name, *old_name;
     char esc_new_name[PATH_MAX * 2], esc_old_name[PATH_MAX * 2];
     char sql[SQL_MAX];
+
+    struct stat to_st;
+
+    
+    if (query_getattr(mysql, to, &to_st) != -ENOENT) {
+        if (S_ISDIR(to_st.st_mode))
+            return -EEXIST;
+    } else {
+        to_st.st_ino = 0;
+    }
 
     inode = query_inode(mysql, from);
 
@@ -1093,6 +1138,19 @@ int query_rename(MYSQL *mysql, const char *from, const char *to)
         log_printf(LOG_ERROR, "Error: mysql_query()\n");
         log_printf(LOG_ERROR, "mysql_error: %s\n", mysql_error(mysql));
         return -EIO;
+    }
+
+    if (to_st.st_ino) {
+        snprintf(sql, SQL_MAX, "DELETE FROM tree "
+                 "WHERE inode = %lu and name = '%s' and parent='%ld'",
+                 to_st.st_ino, esc_new_name, parent_to);
+        log_printf(LOG_D_SQL, "sql=%s\n", sql);
+        ret = mysql_query(mysql, sql);
+        if (ret) {
+            log_printf(LOG_ERROR, "Error: mysql_query()\n");
+            log_printf(LOG_ERROR, "mysql_error: %s\n", mysql_error(mysql));
+            return -EIO;
+        }
     }
 
     /*
